requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
cloudpickle==1.1.1
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==4.4.2
dm-sonnet==1.36
dm-tree==0.1.5
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
graph-nets==1.1.0
grpcio==1.32.0
h5py==2.10.0
importlib-metadata==2.0.0
joblib==0.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.2
matplotlib==3.3.2
networkx==2.5
numpy==1.18.0
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.0.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.23.2
scipy==1.5.3
semantic-version==2.8.5
six==1.15.0
sklearn==0.0
tensorboard==1.15.0
tensorflow==1.15.4
tensorflow-estimator==1.15.1
tensorflow-probability==0.8.0
termcolor==1.1.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.1
ffmpeg-python